First approach to page deduplication.
Each revisit to a URL is expected to return a different page than last time.
After having done page analysis however, we check if its contents are still
the same as those of the previous page we got from this URL, and if they
are, we consider it the same page, and deduplicate it by replacing the
contents of the old one with a 'seeInstead' reference to the new one.

There are many things to be done still:
- The comparison is simply checking for equivalence of body.innerText and
  page title. Better sameness detection is a must.
- Sameness could often be determined beforehand if we check the ETag in the
  HTTP header, thus reducing the need for retrospective deduplication.
- The framework for expressing 'fuzzy sameness' is drafted, but not used.
- Visits could possibly be deduplicated too, e.g. not counting a page reload
  as a new visit.
- Probably a lot more...